+++ packages/runtime/container-runtime/src/dataStoreContext.ts
@@ -58,12 +58,14 @@ import {
     SummarizeInternalFn,
     ITelemetryContext,
 } from "@fluidframework/runtime-definitions";
-import { addBlobToSummary, convertSummaryTreeToITree } from "@fluidframework/runtime-utils";
+import {
+    addBlobToSummary,
+    convertSummaryTreeToITree,
+    packagePathToTelemetryProperty,
+} from "@fluidframework/runtime-utils";
 import {
     ChildLogger,
-    generateStack,
     LoggingError,
-    packagePathToTelemetryProperty,
     TelemetryDataTag,
     ThresholdCounter,
 } from "@fluidframework/telemetry-utils";
@@ -762,19 +764,18 @@ export abstract class FluidDataStoreContext extends TypedEventEmitter<IFluidData
 
         if (checkTombstone && this.tombstoned) {
             const messageString = `Context is tombstoned! Call site [${callSite}]`;
-
-            this.subLogger.sendErrorEvent({
-                eventName: "Tombstone:DataStore_Changed",
-                message: messageString,
-                callSite,
-                pkg: packagePathToTelemetryProperty(this.pkg),
-                stack: generateStack(),
-            });
-
-            throw new DataCorruptionError(messageString, {
+            const error = new DataCorruptionError(messageString, {
                 errorMessage: messageString,
                 ...safeTelemetryProps,
             });
+
+            this.subLogger.sendErrorEvent({
+                eventName: "Tombstone_DataStore_Changed",
+                callSite,
+                pkg: packagePathToTelemetryProperty(this.pkg),
+            }, error);
+
+            throw error;
         }
     }
