+++ packages/runtime/container-runtime/src/dataStores.ts
@@ -34,6 +34,7 @@ import {
     create404Response,
     createResponseError,
     responseToException,
+    packagePathToTelemetryProperty,
     SummaryTreeBuilder,
 } from "@fluidframework/runtime-utils";
 import { ChildLogger, LoggingError, TelemetryDataTag } from "@fluidframework/telemetry-utils";
@@ -429,14 +430,16 @@ export class DataStores implements IDisposable {
         }
 
         if (context.tombstoned) {
+            const error = responseToException(createResponseError(404, "Datastore removed by gc", request), request);
             // Note: if a user writes a request to look like it's viaHandle, we will also send this telemetry event
             this.logger.sendErrorEvent({
-                eventName: "Tombstone:DataStore_Requested",
+                eventName: "Tombstone_DataStore_Requested",
                 url: request.url,
+                pkg: packagePathToTelemetryProperty(context.isLoaded ? context.packagePath : undefined),
                 viaHandle,
-            });
+            }, error);
             // The requested data store is removed by gc. Throw a 404 gc response exception.
-            throw responseToException(createResponseError(404, "Datastore removed by gc", request), request);
+            throw error;
         }
 
         return context;
