+++ packages/runtime/runtime-utils/src/index.ts
@@ -10,6 +10,7 @@ export {
 	createResponseError,
 	exceptionToResponse,
 	Factory,
+    packagePathToTelemetryProperty,
 	requestFluidObject,
 	responseToException,
 } from "./dataStoreHelpers";
