+++ packages/utils/telemetry-utils/src/index.ts
@@ -24,7 +24,6 @@ export {
 	LoggingError,
 	NORMALIZED_ERROR_TYPE,
 	normalizeError,
-    packagePathToTelemetryProperty,
 	wrapError,
 	wrapErrorAndLog,
 } from "./errorLogging";
